Added filterPrefixesOf method to RadixTree
filterPrefixesOf takes a value of the key type and returns a tree
containing the nodes on the path from root to the last matching node
for the supplied key.  No attempt is made to match the remainder of
the key.

This operation is useful for finding all keys represented in the
tree that are prefixes of the search key. For example, when looking
for all filesystem or URI paths in a set of paths that are parents
of a given path.

Note that this operation does not perform the usual path compression
as it traverses the tree.  Methods such as filterPrefix will return
a path-compressed version of the root of the returned tree, as the
individual children of that tree are what is desired.  Since this method
traverses the subtree that it must eventually return, it can't compress
it along the way.